Develop branch init
Develop branch will contain a complete es6 rewrite.

i’d like to add an async-queue, so that multiple commands can’t be sent
to AE at once (ae can’t handle that)

i’d also like to add an option where you can send commands ensuring
that the render engine is being used, instead of just the ui engine.

i’d also like to set up proper testing with mocha.
