+++ .babelrc
@@ -1,0 +1,4 @@
+
+{
+	"presets": ["es2015-node"]
+}
