+++ .languagebabel
@@ -1,0 +1,14 @@
+{
+  "babelMapsPath":                    "null",
+  "babelMapsAddUrl":                  false,
+  "babelSourcePath":                  "src",
+  "babelTranspilePath":               "lib",
+  "createMap":                        false,
+  "createTargetDirectories":          true,
+  "createTranspiledCode":             true,
+  "disableWhenNoBabelrcFileInPath":   true,
+  "projectRoot":                      true,
+  "suppressSourcePathMessages":       true,
+  "suppressTranspileOnSaveMessages":  true,
+  "transpileOnSave":                  true
+}
