+++ lib-old/includes/console.js
