+++ lib-old/platform-mac.js
