+++ lib-old/platform-win.js
