+++ lib/command.js
@@ -1,171 +1,54 @@
 'use strict';
 
-/*******************************************************************/
-// DEPENDENCIES
-/*******************************************************************/
-const fs =     require('fs'),
-      path = 	 require('path'),
-      is = 		 require('is-explicit'),
-      uglify = require('uglify-js'),
-      babel =  require('babel-core');
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
 
-//don't require ae here, as the circular dependencies will result in a {}
-//until initial requires are complete.
-let ae = null;
+var _fs = require('fs');
 
-/*******************************************************************/
-// CONSTANTS
-/*******************************************************************/
+var _fs2 = _interopRequireDefault(_fs);
+
+var _path = require('path');
+
+var _path2 = _interopRequireDefault(_path);
+
+var _isExplicit = require('is-explicit');
+
+var _isExplicit2 = _interopRequireDefault(_isExplicit);
+
+var _uglifyJs = require('uglify-js');
+
+var _uglifyJs2 = _interopRequireDefault(_uglifyJs);
+
+var _babelCore = require('babel-core');
+
+var _babelCore2 = _interopRequireDefault(_babelCore);
+
+var _index = require('./index');
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+/******************************************************************************/
+// Command Specific Configs
+/******************************************************************************/
 
 const BabelOptions = {
-	presets: [
-		require('babel-preset-es2015')
-	],
-	plugins: [
-		require('babel-plugin-transform-es3-member-expression-literals'),
-		require('babel-plugin-transform-es3-property-literals'),
-		require('babel-plugin-transform-es5-property-mutators')
-	],
-	sourceRoot: __dirname
+  presets: [require('babel-preset-es2015')],
+  plugins: [require('babel-plugin-transform-es3-member-expression-literals'), require('babel-plugin-transform-es3-property-literals'), require('babel-plugin-transform-es5-property-mutators')],
+  sourceRoot: __dirname
 };
 
-//These compression options are to satisfy After Effects old javascript engine.
 const CompressionOptions = {
-	conditionals: false
+  conditionals: false
 };
 
-/*******************************************************************/
-// HELPERS
-/*******************************************************************/
+/******************************************************************************/
+// Exports
+/******************************************************************************/
 
-function stringify() { /* private dangler */
-	let err_open = '', err_close = '';
-	if (this.options.errorHandling) {
-		err_open  = '\napp.beginSuppressDialogs();\ntry {\n\n';
-		err_close = '\n} catch (err) {\n' +
-		(is(this.result_file,String) ? '\n\t$.result = err instanceof Error ? err : new Error(err);\n' : '') +
-		'\n}\napp.endSuppressDialogs(false);';
-	}
+class Command {
 
-	let args = '';
-	if (is(this.arguments, Array) && this.arguments.length > 0)
-		args =  ',' + JSON.stringify(this.arguments);
+  constructor(func, options) {}
 
-	let res_get = '', res_write = '';
-	if (is(this.result_file, String)) {
-		res_get = '$.result = ';
-		res_write = '\n\n'+
-		'(function(result){\n'+
-		'\tvar file = File(Folder.temp.absoluteURI+"/' + this.result_file + '");\n'+
-		'\tfile.open("w");\n'+
-		'\tvar str = "module.exports = " + {\n'+
-		'\t\tlogs: typeof console === "object" && typeof console.clearCache === "function" ? console.clearCache() : null,\n'+
-		'\t\treturned: result\n'+
-		'\t}.toSource();\n' +
-		'\tfile.write(str);\n' +
-		'\tfile.close();\n' +
-		'\tdelete $.result;\n'+
-		'})($.result);';
-	}
-
-	let inc = this.includes.join(';');
-	let main = err_open + res_get + this.code + '.apply(this' + args + ');\n' + err_close + res_write;
-
-	// if (this.options.minify)
-	// 	main = main.replace(/\t/g, ' ').replace(/\n/g, ' ').replace(/  /g,'');
-
-	return inc + main;
 }
-
-function babelify(code) {
-	return babel.transform(code, BabelOptions).code;
-}
-
-function minify(code)
-{
-	return uglify.minify(code, {fromString: true, compress: CompressionOptions}).code;
-}
-
-function codify_includes(includes, do_minify) {
-	if (!is(includes, Array)) return [];
-
-	return includes.map(file => {
-		if (!is(file, String)) return null;
-		file = path.resolve(file);
-		try {
-			fs.accessSync(file, fs.R_OK);
-		} catch (err) {
-			throw new Error(`Cannot include file at path ${file}, non-existent or can't be read.`);
-		}
-
-		let code = fs.readFileSync(file, {encoding: 'utf-8'});
-
-		if (path.extname(file) !== '.jsx') {
-			code = babelify(code);
-			if (do_minify)
-				code = minify(code);
-		}
-
-		return code;
-
-	}).filter(inc => is(inc,String));
-}
-
-function execute_args(command, input_args) {
-	let args = [command];
-	for (let i = 0; i < input_args.length; i++) args.push(input_args[i]);
-
-	return args;
-}
-
-/*******************************************************************/
-// CODIFICATION
-/*******************************************************************/
-
-module.exports = class Command {
-
-	constructor(func, options) {
-
-		if (ae === null)
-			ae = require('../index');
-
-		if (!is(func, Function))
-			throw new Error('Commands must be constructed with a function argument.');
-
-		if (!is(options, Object)) options = {};
-
-		this.options = Object.assign({}, ae.options, options);
-		//Pass function as an expression otherwise it breaks when isolated
-		this.code = babelify(`(${func.toString()})`);
-
-		//Isolate babelified code to just the funciton expression (remove 'use strict' and final ';')
-		this.code = this.code.substring(this.code.indexOf('(function'), this.code.length - 1);
-
-		if (this.options.minify) {
-			//minify will return no code if its just a function expression that doesn't get executed
-			//so we have to trick it and then we-wrap the minified code into a function expression.
-			this.code = minify('var temp = ' + this.code + '()');
-			this.code = '(' + this.code.substring(9, this.code.length - 3) + ')';
-		}
-
-		this.includes = codify_includes(this.options.includes, this.options.minify);
-
-		Object.freeze(this.options);
-		Object.freeze(this.options.includes);
-
-	}
-
-	toString() {
-		return stringify.call(this);
-	}
-
-	execute() {
-		let args = execute_args(this, arguments);
-		return ae.execute.apply(null, args);
-	}
-
-	executeSync() {
-		let args = execute_args(this, arguments);
-		return ae.executeSync.apply(null, args);
-	}
-};
+exports.default = Command;
